Media service controler

+++ PopcornPals-Server/src/main/java/bg/fmi/popcornpals/controller/StudioController.java
@@ -73,6 +73,10 @@ public class StudioController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteStudioById(@PathVariable Long id) {
+        Studio existingStudio = studioService.getStudios(id, null).get(0);
+        if (existingStudio == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         studioService.deleteStudioById(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
